server/requirements.txt
asgiref
Django
django-redis
djangorestframework
dnspython
mongoengine
psycopg
pymongo
python-dotenv
redis
sqlparse
